% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R, R/mlwinfitMCMC.R
\docType{methods}
\name{coef,mlwinfitIGLS-method}
\alias{coef,mlwinfitIGLS-method}
\alias{coef,mlwinfitMCMC-method}
\alias{coefficients,mlwinfitIGLS-method}
\alias{coefficients,mlwinfitMCMC-method}
\title{Extract the coefficient vector from "mlwinfitIGLS" objects}
\usage{
\S4method{coef}{mlwinfitIGLS}(object, ...)

\S4method{coefficients}{mlwinfitIGLS}(object, ...)

\S4method{coef}{mlwinfitMCMC}(object, ...)

\S4method{coefficients}{mlwinfitMCMC}(object, ...)
}
\description{
Extract the coefficient vector from "mlwinfitIGLS" objects

Extract the coefficient vector from "mlwinfitMCMC" objects
}

