% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{coef,mlwinfitIGLS-method}
\alias{coef,mlwinfitIGLS-method}
\alias{coefficients,mlwinfitIGLS-method}
\title{Extract the coefficient vector from "mlwinfitIGLS" objects}
\usage{
\S4method{coef}{mlwinfitIGLS}(object, ...)

\S4method{coefficients}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object}

\item{...}{Other arguments}
}
\description{
Extract the coefficient vector from "mlwinfitIGLS" objects
}
\seealso{
\code{\link[stats]{coef}}
}
