% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{coef.mlwinfitIGLS}
\alias{coef.mlwinfitIGLS}
\title{Extract the coefficient vector from "mlwinfitIGLS" objects}
\usage{
\method{coef}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object}

\item{...}{Other arguments}
}
\description{
Extract the coefficient vector from "mlwinfitIGLS" objects
}
\seealso{
\code{\link[stats]{coef}}
}
