% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{coef.mlwinfitMCMC}
\alias{coef.mlwinfitMCMC}
\title{Extract the coefficient vector from "mlwinfitMCMC" objects}
\usage{
\method{coef}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object}

\item{...}{Other arguments}
}
\description{
Extract the coefficient vector from "mlwinfitMCMC" objects
}
\seealso{
\code{\link[stats]{coef}}
}
