% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{deviance,mlwinfitIGLS-method}
\alias{deviance,mlwinfitIGLS-method}
\title{Returns the deviance from "mlwinfitIGLS" objects.}
\usage{
\S4method{deviance}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object}

\item{...}{Other arguments}
}
\description{
Returns the deviance from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{deviance}}
}

