% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{deviance.mlwinfitIGLS}
\alias{deviance.mlwinfitIGLS}
\title{Returns the deviance from "mlwinfitIGLS" objects.}
\usage{
\method{deviance}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object}

\item{...}{Other arguments}
}
\description{
Returns the deviance from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{deviance}}
}
