% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R, R/mlwinfitMCMC.R
\docType{methods}
\name{df.residual,mlwinfitIGLS-method}
\alias{df.residual,mlwinfitIGLS-method}
\alias{df.residual,mlwinfitMCMC-method}
\title{Returns the residual degrees-of-freedom extracted from "mlwinfitIGLS" objects.}
\usage{
\S4method{df.residual}{mlwinfitIGLS}(object, ...)

\S4method{df.residual}{mlwinfitMCMC}(object, ...)
}
\description{
Returns the residual degrees-of-freedom extracted from "mlwinfitIGLS" objects.

Returns the residual degrees-of-freedom extracted from "mlwinfitMCMC" objects.
}

