% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{df.residual,mlwinfitMCMC-method}
\alias{df.residual,mlwinfitMCMC-method}
\title{Returns the residual degrees-of-freedom extracted from "mlwinfitMCMC" objects.}
\usage{
\S4method{df.residual}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object.}

\item{...}{Other arguments}
}
\description{
Returns the residual degrees-of-freedom extracted from "mlwinfitMCMC" objects.
}
\seealso{
\code{\link[stats]{nobs}}, \code{\link[stats]{coef}}
}

