% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{df.residual.mlwinfitIGLS}
\alias{df.residual.mlwinfitIGLS}
\title{Returns the residual degrees-of-freedom extracted from "mlwinfitIGLS" objects.}
\usage{
\method{df.residual}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object.}

\item{...}{Other arguments}
}
\description{
Returns the residual degrees-of-freedom extracted from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{nobs}}, \code{\link[stats]{coef}}
}
