% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{[,mlwinfitIGLS,ANY,ANY,ANY-method}
\alias{[,mlwinfitIGLS,ANY,ANY,ANY-method}
\alias{[<-,mlwinfitIGLS,ANY,ANY,ANY-method}
\alias{[[,mlwinfitIGLS-method}
\alias{[[<-,mlwinfitIGLS-method}
\title{Extract or Replace parts of "mlwinfitIGLS" objects}
\usage{
\S4method{[}{mlwinfitIGLS,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{mlwinfitIGLS,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[[}{mlwinfitIGLS}(x, i, j, drop)

\S4method{[[}{mlwinfitIGLS}(x, i, j) <- value
}
\description{
Extract or Replace parts of "mlwinfitIGLS" objects
}

