% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{extract,mlwinfitIGLS-method}
\alias{extract,mlwinfitIGLS-method}
\title{Extract coefficients and GOF measures from a statistical object (texreg package).}
\usage{
\S4method{extract}{mlwinfitIGLS}(model, include.nobs = TRUE,
  include.loglik = TRUE, include.deviance = TRUE, ...)
}
\arguments{
\item{model}{An \code{\link{mlwinfitIGLS-class}} model.}

\item{include.nobs}{should the number of observations be reported?}

\item{include.loglik}{should the log-likelihood be reported?}

\item{include.deviance}{should the deviance be reported?}

\item{...}{Other arguments.}
}
\description{
Extract coefficients and GOF measures from a statistical object (texreg package).
}
\seealso{
\code{\link[texreg]{extract}}
}

