% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{extract,mlwinfitMCMC-method}
\alias{extract,mlwinfitMCMC-method}
\title{Extract coefficients and GOF measures from a statistical object.}
\usage{
\S4method{extract}{mlwinfitMCMC}(model, include.nobs = TRUE,
  include.dbar = TRUE, include.dthetabar = TRUE, include.pd = TRUE,
  include.dic = TRUE, ...)
}
\arguments{
\item{model}{An \code{\link{mlwinfitMCMC-class}} model.}

\item{include.nobs}{should the number of observations be reported?}

\item{include.dbar}{should the Dbar be reported?}

\item{include.dthetabar}{should the D(thetabar) be reported?}

\item{include.pd}{should the pD be reported?}

\item{include.dic}{should the DIC be reported?}

\item{...}{Other arguments.}
}
\description{
Extract coefficients and GOF measures from a statistical object.
}
\seealso{
\code{\link[texreg]{extract}}
}
