% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{fitted.mlwinfitIGLS}
\alias{fitted.mlwinfitIGLS}
\title{Returns the fitted values from "mlwinfitIGLS" objects.}
\usage{
\method{fitted}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object.}

\item{...}{Other arguments.}
}
\description{
Returns the fitted values from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{fitted.values}}
}
