% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{fitted.mlwinfitMCMC}
\alias{fitted.mlwinfitMCMC}
\title{Returns the fitted values from "mlwinfitMCMC" objects.}
\usage{
\method{fitted}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object.}

\item{...}{Other arguments}
}
\description{
Returns the fitted values from "mlwinfitMCMC" objects.
}
\seealso{
\code{\link[stats]{fitted.values}}
}
