% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{formula,mlwinfitIGLS-method}
\alias{formula,mlwinfitIGLS-method}
\title{"mlwinfitIGLS" model formula}
\usage{
\S4method{formula}{mlwinfitIGLS}(x, env = parent.frame(), ...)
}
\arguments{
\item{x}{See \code{\link[stats]{formula}}}

\item{env}{See \code{\link[stats]{formula}}}

\item{...}{Other arguments; see \code{\link[stats]{formula}}}
}
\description{
"mlwinfitIGLS" model formula
}

