% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{formula,mlwinfitMCMC-method}
\alias{formula,mlwinfitMCMC-method}
\title{"mlwinfitMCMC" model formula}
\usage{
\S4method{formula}{mlwinfitMCMC}(x, env = parent.frame(), ...)
}
\arguments{
\item{x}{See \code{\link[stats]{formula}}}

\item{env}{See \code{\link[stats]{formula}}}

\item{...}{Other arguments; see \code{\link[stats]{formula}}}
}
\description{
"mlwinfitMCMC" model formula
}

