% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{formula.mlwinfitIGLS}
\alias{formula.mlwinfitIGLS}
\title{"mlwinfitIGLS" model formula}
\usage{
\method{formula}{mlwinfitIGLS}(x, env = parent.frame(), ...)
}
\arguments{
\item{x}{See \code{\link[stats]{formula}}}

\item{env}{See \code{\link[stats]{formula}}}

\item{...}{Other arguments; see \code{\link[stats]{formula}}}
}
\description{
"mlwinfitIGLS" model formula
}
