% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{formula.mlwinfitMCMC}
\alias{formula.mlwinfitMCMC}
\title{"mlwinfitMCMC" model formula}
\usage{
\method{formula}{mlwinfitMCMC}(x, env = parent.frame(), ...)
}
\arguments{
\item{x}{See \code{\link[stats]{formula}}}

\item{env}{See \code{\link[stats]{formula}}}

\item{...}{Other arguments; see \code{\link[stats]{formula}}}
}
\description{
"mlwinfitMCMC" model formula
}
