% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R, R/mlwinfitMCMC.R
\docType{methods}
\name{formula,mlwinfitIGLS-method}
\alias{formula,mlwinfitIGLS-method}
\alias{formula,mlwinfitMCMC-method}
\title{"mlwinfitIGLS" model formula}
\usage{
\S4method{formula}{mlwinfitIGLS}(x, env = parent.frame(), ...)

\S4method{formula}{mlwinfitMCMC}(x, env = parent.frame(), ...)
}
\description{
"mlwinfitIGLS" model formula

"mlwinfitMCMC" model formula
}

