% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{getSummary,mlwinfitMCMC-method}
\alias{getSummary,mlwinfitMCMC-method}
\title{Extract coefficients and GOF measures from a statistical object (memisc package).}
\usage{
\S4method{getSummary}{mlwinfitMCMC}(obj, alpha = 0.05, ...)
}
\arguments{
\item{obj}{An \code{\link{mlwinfitIGLS-class}} model.}

\item{alpha}{level of the confidence intervals; their coverage should be 1-alpha/2}

\item{...}{Other arguments.}
}
\description{
Extract coefficients and GOF measures from a statistical object (memisc package).
}
\seealso{
\code{\link[memisc]{getSummary}}
}

