% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{glance.mlwinfitIGLS}
\alias{glance.mlwinfitIGLS}
\title{Extract GOF measures from a statistical object (broom package).}
\usage{
\method{glance}{mlwinfitIGLS}(x, ...)
}
\arguments{
\item{x}{An \code{\link{mlwinfitIGLS-class}} model.}

\item{...}{Other arguments.}
}
\description{
Extract GOF measures from a statistical object (broom package).
}
\seealso{
\code{\link[generics]{glance}}
}
