% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{glance.mlwinfitMCMC}
\alias{glance.mlwinfitMCMC}
\title{Extract GOF measures from a statistical object (broom package).}
\usage{
\method{glance}{mlwinfitMCMC}(x, ...)
}
\arguments{
\item{x}{An \code{\link{mlwinfitMCMC-class}} model.}

\item{...}{Other arguments.}
}
\description{
Extract GOF measures from a statistical object (broom package).
}
\seealso{
\code{\link[generics]{glance}}
}
