% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{logLik,mlwinfitIGLS-method}
\alias{logLik,mlwinfitIGLS-method}
\title{Returns the log-likelihood from "mlwinfitIGLS" objects.}
\usage{
\S4method{logLik}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object.}

\item{...}{Other arguments.}
}
\description{
Returns the log-likelihood from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{logLik}}
}

