% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{logLik.mlwinfitIGLS}
\alias{logLik.mlwinfitIGLS}
\title{Returns the log-likelihood from "mlwinfitIGLS" objects.}
\usage{
\method{logLik}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object.}

\item{...}{Other arguments.}
}
\description{
Returns the log-likelihood from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{logLik}}
}
