% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R, R/mlwinfitMCMC.R
\docType{methods}
\name{nobs,mlwinfitIGLS-method}
\alias{nobs,mlwinfitIGLS-method}
\alias{nobs,mlwinfitMCMC-method}
\title{Returns the number of used observations from "mlwinfitIGLS" objects.}
\usage{
\S4method{nobs}{mlwinfitIGLS}(object, ...)

\S4method{nobs}{mlwinfitMCMC}(object, ...)
}
\description{
Returns the number of used observations from "mlwinfitIGLS" objects.

Returns the number of used observations from "mlwinfitMCMC" objects.
}

