% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{nobs,mlwinfitIGLS-method}
\alias{nobs,mlwinfitIGLS-method}
\title{Returns the number of used observations from "mlwinfitIGLS" objects.}
\usage{
\S4method{nobs}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object.}

\item{...}{Other arguments.}
}
\description{
Returns the number of used observations from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{nobs}}
}

