% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{nobs,mlwinfitMCMC-method}
\alias{nobs,mlwinfitMCMC-method}
\title{Returns the number of used observations from "mlwinfitMCMC" objects.}
\usage{
\S4method{nobs}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object.}

\item{...}{Other arguments.}
}
\description{
Returns the number of used observations from "mlwinfitMCMC" objects.
}
\seealso{
\code{\link[stats]{nobs}}
}

