% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{nobs.mlwinfitIGLS}
\alias{nobs.mlwinfitIGLS}
\title{Returns the number of used observations from "mlwinfitIGLS" objects.}
\usage{
\method{nobs}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object.}

\item{...}{Other arguments.}
}
\description{
Returns the number of used observations from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{nobs}}
}
