% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R, R/mlwinfitMCMC.R
\docType{methods}
\name{fitted,mlwinfitIGLS-method}
\alias{fitted,mlwinfitIGLS-method}
\alias{fitted,mlwinfitMCMC-method}
\alias{fitted.values,mlwinfitIGLS-method}
\alias{fitted.values,mlwinfitMCMC-method}
\alias{predict,mlwinfitIGLS-method}
\alias{predict,mlwinfitMCMC-method}
\title{Returns the predicted data from "mlwinfitIGLS" objects.}
\usage{
\S4method{fitted}{mlwinfitIGLS}(object, ...)

\S4method{fitted.values}{mlwinfitIGLS}(object, ...)

\S4method{predict}{mlwinfitIGLS}(object, newdata = NULL, params = NULL,
  type = "link", se.fit = FALSE, terms = NULL, ...)

\S4method{fitted}{mlwinfitMCMC}(object, ...)

\S4method{fitted.values}{mlwinfitMCMC}(object, ...)

\S4method{predict}{mlwinfitMCMC}(object, newdata = NULL, params = NULL,
  type = "link", se.fit = FALSE, terms = NULL, ...)
}
\description{
Returns the predicted data from "mlwinfitIGLS" objects.

Returns the predicted data from "mlwinfitMCMC" objects.
}

