% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{predict,mlwinfitIGLS-method}
\alias{predict,mlwinfitIGLS-method}
\title{Returns the predicted data from "mlwinfitIGLS" objects.}
\usage{
\S4method{predict}{mlwinfitIGLS}(object, newdata = NULL, params = NULL,
  type = "link", se.fit = FALSE, terms = NULL, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object.}

\item{newdata}{data frame for which to evaluate predictions}

\item{params}{a character vector specifying the parameters to use in evaluating predictions.
If \code{NULL}, \code{names(object[["FP"]])} is used by default.}

\item{type}{when this has the value \code{"link"} (default) the linear predictor is returned.
When \code{type="terms"} each component of the linear predictor is returned seperately. When \code{type="response"} predictions on the scale of the response are returned.}

\item{se.fit}{logical. When this is \code{TRUE} (not default) standard error estimates are returned for each prediction.}

\item{terms}{if \code{type="terms"}, which terms (default is all terms), a character vector.}

\item{...}{Other arguments.}
}
\description{
Returns the predicted data from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{predict}}
}
