% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{predict,mlwinfitMCMC-method}
\alias{predict,mlwinfitMCMC-method}
\title{Returns the predicted data from "mlwinfitMCMC" objects.}
\usage{
\S4method{predict}{mlwinfitMCMC}(object, newdata = NULL, params = NULL,
  type = "link", se.fit = FALSE, terms = NULL, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object.}

\item{newdata}{data frame for which to evaluate predictions}

\item{params}{a character vector specifying the parameters to use in evaluating predictions. If \code{NULL}, \code{names(object[["FP"]])} is used by default.}

\item{type}{when this has the value \code{"link"} (default) the linear predictor is returned.
When \code{type="terms"} each component of the linear predictor is returned seperately. When \code{type="response"} predictions on the scale of the response are returned.}

\item{se.fit}{logical. When this is \code{TRUE} (not default) standard error estimates are returned for each prediction.}

\item{terms}{if \code{type="terms"}, which terms (default is all terms), a character vector.}

\item{...}{Other arguments}
}
\description{
Returns the predicted data from "mlwinfitMCMC" objects.
}
\seealso{
\code{\link[stats]{predict}}
}

