\docType{data}
\name{reading1}
\alias{reading1}
\title{.}

\format{A data frame with 407 observations on the following 13 variables:
  \describe{
    \item{\code{ID}}{Unique pupil identifying code.}
    \item{\code{AGE1}}{Age at occasion 1.}
    \item{\code{READ1}}{Reading score at occasion 1.}
    \item{\code{AGE2}}{Age at occasion 2.}
    \item{\code{READ2}}{Reading score at occasion 2.}
    \item{\code{AGE3}}{Age at occasion 3.}
    \item{\code{READ3}}{Reading score at occasion 3.}
    \item{\code{AGE4}}{Age at occasion 4.}
    \item{\code{READ4}}{Reading score at occasion 4.}
    \item{\code{AGE5}}{Age at occasion 5.}
    \item{\code{READ5}}{Reading score at occasion 5.}
    \item{\code{AGE6}}{Age at occasion 6.}
    \item{\code{READ6}}{Reading score at occasion 6.}

}}
\source{
Rasbash, J., Browne, W. J., Healy, M., Cameron, B., Charlton, C. M. J. (2013) MLwiN v2.27. University of Bristol: Centre for Multilevel Modelling.
Tizard, B., Blatchford, P., Burke, J. & Farquhar, C. (1988). Young children at school in the inner city. Hove, Sussex: Lawrence Erlbaum.
}

\description{
  Reading score data for 407 pupils across 6 occasions.}
\details{
  The \code{reading1} dataset is one of the sample datasets provided with the multilevel-modelling software package MLwiN (Rasbash et al., 2013), and was analysed in Tizard et al. (1988).}

\examples{
\dontrun{
#NOTE: Assumes MLwiN path is C:/Program Files (x86)/MLwiN v2.30/
#...so please change relevant line if different
#if using R2MLwiN via WINE, the path may look like 
#options(MLwiN_path="/home/USERNAME/.wine/drive_c/Program Files (x86)/MLwiN v2.30/")

data(reading1)




}
}
\keyword{datasets}

