% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R, R/mlwinfitMCMC.R
\docType{methods}
\name{residuals,mlwinfitIGLS-method}
\alias{resid,mlwinfitIGLS-method}
\alias{resid,mlwinfitMCMC-method}
\alias{residuals,mlwinfitIGLS-method}
\alias{residuals,mlwinfitMCMC-method}
\title{Returns the residual data from "mlwinfitIGLS" objects.}
\usage{
\S4method{residuals}{mlwinfitIGLS}(object, ...)

\S4method{resid}{mlwinfitIGLS}(object, ...)

\S4method{residuals}{mlwinfitMCMC}(object, ...)

\S4method{resid}{mlwinfitMCMC}(object, ...)
}
\description{
Returns the residual data from "mlwinfitIGLS" objects.

Returns the residual data from "mlwinfitMCMC" objects.
}

