% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{residuals,mlwinfitIGLS-method}
\alias{resid,mlwinfitIGLS-method}
\alias{residuals,mlwinfitIGLS-method}
\title{Returns the residual data from "mlwinfitIGLS" objects.}
\usage{
\S4method{residuals}{mlwinfitIGLS}(object, ...)

\S4method{resid}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object}

\item{...}{Other arguments.}
}
\description{
Returns the residual data from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{residuals}}
}

