% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{residuals,mlwinfitMCMC-method}
\alias{residuals,mlwinfitMCMC-method}
\alias{resid,mlwinfitMCMC-method}
\title{Returns the residual data from "mlwinfitMCMC" objects.}
\usage{
\S4method{residuals}{mlwinfitMCMC}(object, ...)

\S4method{resid}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object}

\item{...}{Other arguments.}
}
\description{
Returns the residual data from "mlwinfitMCMC" objects.
}
\seealso{
\code{\link[stats]{residuals}}
}
