% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{residuals.mlwinfitIGLS}
\alias{residuals.mlwinfitIGLS}
\title{Returns the residual data from "mlwinfitIGLS" objects.}
\usage{
\method{residuals}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object}

\item{...}{Other arguments.}
}
\description{
Returns the residual data from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{residuals}}
}
