% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{residuals.mlwinfitMCMC}
\alias{residuals.mlwinfitMCMC}
\title{Returns the residual data from "mlwinfitMCMC" objects.}
\usage{
\method{residuals}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object}

\item{...}{Other arguments.}
}
\description{
Returns the residual data from "mlwinfitMCMC" objects.
}
\seealso{
\code{\link[stats]{residuals}}
}
