% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{show,mlwinfitIGLS-method}
\alias{show,mlwinfitIGLS-method}
\title{Show objects of class "mlwinfitIGLS"}
\usage{
\S4method{show}{mlwinfitIGLS}(object)
}
\arguments{
\item{object}{an \code{\link{mlwinfitIGLS-class}} object}
}
\description{
Show objects of class "mlwinfitIGLS"
}
\seealso{
\code{\link[stats4]{show-methods}}
}
