% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{show.mlwinfitMCMC}
\alias{show.mlwinfitMCMC}
\title{Summarize "mlwinfitMCMC" objects}
\usage{
\method{show}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{an \code{\link{mlwinfitMCMC-class}} object}

\item{...}{other parameters}
}
\description{
Summarize "mlwinfitMCMC" objects
}
\seealso{
\code{\link[methods]{show}}
}
