% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{summary,mlwinfitIGLS-method}
\alias{print,mlwinfitIGLS-method}
\alias{show,mlwinfitIGLS-method}
\alias{summary,mlwinfitIGLS-method}
\title{Summarize "mlwinfitIGLS" objects}
\usage{
\S4method{summary}{mlwinfitIGLS}(object, ...)

\S4method{print}{mlwinfitIGLS}(x, digits = max(3, getOption("digits") - 2),
  signif.stars = getOption("show.signif.stars"), ...)

\S4method{show}{mlwinfitIGLS}(object)
}
\description{
Summarize "mlwinfitIGLS" objects
}

