% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{summary,mlwinfitMCMC-method}
\alias{print,mlwinfitMCMC-method}
\alias{show,mlwinfitMCMC-method}
\alias{summary,mlwinfitMCMC-method}
\title{Summarize "mlwinfitMCMC" objects}
\usage{
\S4method{summary}{mlwinfitMCMC}(object, ...)

\S4method{print}{mlwinfitMCMC}(x, digits = max(3, getOption("digits") - 2),
  signif.stars = getOption("show.signif.stars"), z.ratio = TRUE, ...)

\S4method{show}{mlwinfitMCMC}(object)
}
\description{
Summarize "mlwinfitMCMC" objects
}

