% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{summary,mlwinfitIGLS-method}
\alias{summary,mlwinfitIGLS-method}
\title{Summarize "mlwinfitIGLS" objects}
\usage{
\S4method{summary}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object, x}{an \code{\link{mlwinfitIGLS-class}} object}

\item{...}{other parameters}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If TRUE, 'significance stars' are printed for each coefficient.}
}
\description{
Summarize "mlwinfitIGLS" objects
}
\seealso{
\code{\link[stats4]{summary-methods}}
}
