% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{summary,mlwinfitMCMC-method}
\alias{summary,mlwinfitMCMC-method}
\title{Summarize "mlwinfitMCMC" objects}
\usage{
\S4method{summary}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object, x}{an \code{\link{mlwinfitMCMC-class}} object}

\item{...}{other parameters}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If TRUE, 'significance stars' are printed for each coefficient.}

\item{z.ratio}{logical. If TRUE, z-ratio values are displayed for each coefficient.}
}
\description{
Summarize "mlwinfitMCMC" objects
}
\seealso{
\code{\link[stats4]{summary-methods}}
}
