% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{summary.mlwinfitMCMC}
\alias{summary.mlwinfitMCMC}
\title{Summarize "mlwinfitMCMC" objects}
\usage{
\method{summary}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{an \code{\link{mlwinfitMCMC-class}} object}

\item{...}{other parameters}
}
\description{
Summarize "mlwinfitMCMC" objects
}
