% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{tidy.mlwinfitIGLS}
\alias{tidy.mlwinfitIGLS}
\title{Summarises information about the components of a model from a statistical object (broom package).}
\usage{
\method{tidy}{mlwinfitIGLS}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An \code{\link{mlwinfitIGLS-class}} model.}

\item{conf.int}{should the confidence interval be included?}

\item{conf.level}{confidence interval level}

\item{...}{Other arguments.}
}
\description{
Summarises information about the components of a model from a statistical object (broom package).
}
\seealso{
\code{\link[generics]{tidy}}
}
