% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{update,mlwinfitIGLS-method}
\alias{update,mlwinfitIGLS-method}
\title{Update "mlwinfitIGLS" objects}
\usage{
\S4method{update}{mlwinfitIGLS}(object, Formula., levID., estoptions., ...,
  keep.order = TRUE, evaluate = TRUE)
}
\description{
Update "mlwinfitIGLS" objects
}

