% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{update,mlwinfitMCMC-method}
\alias{update,mlwinfitMCMC-method}
\title{Update "mlwinfitMCMC" objects}
\usage{
\S4method{update}{mlwinfitMCMC}(object, Formula., levID., estoptions., ...,
  keep.order = TRUE, evaluate = TRUE)
}
\description{
Update "mlwinfitMCMC" objects
}

