% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R, R/mlwinfitMCMC.R
\docType{methods}
\name{vcov,mlwinfitIGLS-method}
\alias{vcov,mlwinfitIGLS-method}
\alias{vcov,mlwinfitMCMC-method}
\title{Extract the approximate variance-covariance matrix from "mlwinfitIGLS" objects}
\usage{
\S4method{vcov}{mlwinfitIGLS}(object, ...)

\S4method{vcov}{mlwinfitMCMC}(object, ...)
}
\description{
Extract the approximate variance-covariance matrix from "mlwinfitIGLS" objects

Extract the approximate variance-covariance matrix from "mlwinfitMCMC" objects
}

