% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{vcov,mlwinfitIGLS-method}
\alias{vcov,mlwinfitIGLS-method}
\title{Extract the approximate variance-covariance matrix from "mlwinfitIGLS" objects}
\usage{
\S4method{vcov}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object}

\item{...}{Other arguments}
}
\description{
Extract the approximate variance-covariance matrix from "mlwinfitIGLS" objects
}
\seealso{
\code{\link[stats]{vcov}}
}

