% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{vcov,mlwinfitMCMC-method}
\alias{vcov,mlwinfitMCMC-method}
\title{Extract the approximate variance-covariance matrix from "mlwinfitMCMC" objects}
\usage{
\S4method{vcov}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object}

\item{...}{Other arguments}
}
\description{
Extract the approximate variance-covariance matrix from "mlwinfitMCMC" objects
}
\seealso{
\code{\link[stats]{vcov}}
}

