% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{vcov.mlwinfitIGLS}
\alias{vcov.mlwinfitIGLS}
\title{Extract the approximate variance-covariance matrix from "mlwinfitIGLS" objects}
\usage{
\method{vcov}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object}

\item{...}{Other arguments}
}
\description{
Extract the approximate variance-covariance matrix from "mlwinfitIGLS" objects
}
\seealso{
\code{\link[stats]{vcov}}
}
