% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{vcov.mlwinfitMCMC}
\alias{vcov.mlwinfitMCMC}
\title{Extract the approximate variance-covariance matrix from "mlwinfitMCMC" objects}
\usage{
\method{vcov}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object}

\item{...}{Other arguments}
}
\description{
Extract the approximate variance-covariance matrix from "mlwinfitMCMC" objects
}
\seealso{
\code{\link[stats]{vcov}}
}
