% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{augment.mlwinfitIGLS}
\alias{augment.mlwinfitIGLS}
\title{Augment data frame with information derived from the model fit (broom package).}
\usage{
\method{augment}{mlwinfitIGLS}(x, data = x@frame, newdata = NULL, type.predict, type.residuals, ...)
}
\arguments{
\item{x}{An \code{\link{mlwinfitIGLS-class}} model.}

\item{data}{original data onto which columns should be added}

\item{newdata}{new data to predict on, optional}

\item{type.predict}{Type of prediction to compute}

\item{type.residuals}{Type of residuals to compute}

\item{...}{Other arguments.}
}
\description{
Augment data frame with information derived from the model fit (broom package).
}
\seealso{
\code{\link[generics]{augment}}
}
