% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{[,mlwinfitIGLS,ANY,ANY,ANY-method}
\alias{[,mlwinfitIGLS,ANY,ANY,ANY-method}
\alias{[<-,mlwinfitIGLS,ANY,ANY,ANY-method}
\alias{[[,mlwinfitIGLS-method}
\alias{[[<-,mlwinfitIGLS-method}
\title{Extract or Replace parts of "mlwinfitIGLS" objects}
\usage{
\S4method{[}{mlwinfitIGLS,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{mlwinfitIGLS,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[[}{mlwinfitIGLS}(x, i, j, drop)

\S4method{[[}{mlwinfitIGLS}(x, i, j) <- value
}
\arguments{
\item{x}{data frame}

\item{i, j}{elements to extract or replace. For \code{[} and \code{[[}, these are \code{character}.}

\item{drop}{not used.}

\item{value}{a suitable replacement value.}
}
\description{
Extract or Replace parts of "mlwinfitIGLS" objects
}
