% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{[,mlwinfitMCMC,ANY,ANY,ANY-method}
\alias{[,mlwinfitMCMC,ANY,ANY,ANY-method}
\alias{[<-,mlwinfitMCMC,ANY,ANY,ANY-method}
\alias{[[,mlwinfitMCMC-method}
\alias{[[<-,mlwinfitMCMC-method}
\title{Extract or Replace parts of "mlwinfitMCMC" objects}
\usage{
\S4method{[}{mlwinfitMCMC,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{mlwinfitMCMC,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[[}{mlwinfitMCMC}(x, i, j, drop)

\S4method{[[}{mlwinfitMCMC}(x, i, j) <- value
}
\arguments{
\item{x}{data frame}

\item{i, j}{elements to extract or replace. For \code{[} and \code{[[}, these are \code{character}.}

\item{drop}{not used.}

\item{value}{a suitable replacement value.}
}
\description{
Extract or Replace parts of "mlwinfitMCMC" objects
}

