% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R, R/mlwinfitMCMC.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,mlwinfitIGLS-method}
\alias{[,mlwinfitMCMC-method}
\alias{[<-,mlwinfitIGLS-method}
\alias{[<-,mlwinfitMCMC-method}
\alias{[[}
\alias{[[,mlwinfitIGLS-method}
\alias{[[,mlwinfitMCMC-method}
\title{Extract parts of mlwinfitIGLS}
\usage{
\S4method{[}{mlwinfitIGLS,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[[}{mlwinfitIGLS}(x, i, j, drop)

\S4method{[}{mlwinfitIGLS,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[[}{mlwinfitIGLS}(x, i, j) <- value

\S4method{[}{mlwinfitMCMC,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{mlwinfitMCMC,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[[}{mlwinfitMCMC}(x, i, j, drop)

\S4method{[[}{mlwinfitMCMC}(x, i, j) <- value
}
\description{
Extract parts of mlwinfitIGLS

Extract parts of mlwinfitIGLS

Replace names of mlwinfitIGLS

Replace names of mlwinfitIGLS

This method gets a slot of an instance of the S4 class.

Replace names of mlwinfitMCMC

Extract parts of mlwinfitMCMC

Replace names of mlwinfitMCMC
}

