% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{fitted,mlwinfitIGLS-method}
\alias{fitted,mlwinfitIGLS-method}
\alias{fitted.values,mlwinfitIGLS-method}
\title{Returns the fitted values from "mlwinfitIGLS" objects.}
\usage{
\S4method{fitted}{mlwinfitIGLS}(object, ...)

\S4method{fitted.values}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object.}

\item{...}{Other arguments.}
}
\description{
Returns the fitted values from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{fitted}}
}

