% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{fitted,mlwinfitIGLS-method}
\alias{fitted,mlwinfitIGLS-method}
\alias{fitted.values,mlwinfitIGLS-method}
\alias{predict,mlwinfitIGLS-method}
\title{Returns the predicted data from "mlwinfitIGLS" objects.}
\usage{
\S4method{fitted}{mlwinfitIGLS}(object, ...)

\S4method{fitted.values}{mlwinfitIGLS}(object, ...)

\S4method{predict}{mlwinfitIGLS}(object, newdata = NULL, params = NULL,
  type = "link", se.fit = FALSE, terms = NULL, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitIGLS-class}} object.}

\item{...}{Other arguments.}
}
\description{
Returns the predicted data from "mlwinfitIGLS" objects.
}
\seealso{
\code{\link[stats]{predict}}
}

