% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{fitted,mlwinfitMCMC-method}
\alias{fitted,mlwinfitMCMC-method}
\alias{fitted.values,mlwinfitMCMC-method}
\title{Returns the fitted values from "mlwinfitMCMC" objects.}
\usage{
\S4method{fitted}{mlwinfitMCMC}(object, ...)

\S4method{fitted.values}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mlwinfitMCMC-class}} object.}

\item{...}{Other arguments}
}
\description{
Returns the fitted values from "mlwinfitMCMC" objects.
}
\seealso{
\code{\link[stats]{fitted}}
}

