\docType{data}
\name{height}
\alias{height}
\title{Height data.}
\format{A data frame with 100 observations on the following variable:
  \describe{
    \item{\code{height}}{Heights of 100 adult males measured in centimetres.}
}}
\source{
Rasbash, J., Browne, W. J., Healy, M., Cameron, B., Charlton, C. M. J. (2013) MLwiN v2.27. University of Bristol: Centre for Multilevel Modelling.
}

\description{
  Height data for 100 adult males.}
\details{
  The \code{height} dataset is one of the sample datasets provided with the multilevel-modelling software package MLwiN (Rasbash et al., 2013).}


\examples{
\dontrun{
#NOTE: Assumes MLwiN path is C:/Program Files (x86)/MLwiN v2.30/
#...so please change relevant line if different
#if using R2MLwiN via WINE, the path may look like 
#options(MLwiN_path="/home/USERNAME/.wine/drive_c/Program Files (x86)/MLwiN v2.30/")

data(height)


}
}
\keyword{datasets}

