% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predCurves.r
\name{predCurves}
\alias{predCurves}
\title{Draws predicted curves (lines) using estimates from the fixed part of a
fitted model.}
\usage{
predCurves(object, indata = NULL, xname, group = NULL, legend = TRUE,
  legend.space = "top", legend.ncol = 2, ...)
}
\arguments{
\item{object}{Either an \code{\link{mlwinfitIGLS-class}} or \code{\link{mlwinfitMCMC-class}} object.}

\item{indata}{A data.frame object containing the data. If not specified, data is extracted from
the \code{object}.}

\item{xname}{The name of variable to be plotted.}

\item{group}{A character string or a sequence of length equivalent to rows of data to plot.
\code{group = NULL} by default.}

\item{legend}{A logical value indicating whether a legend for \code{group}
is to be added.}

\item{legend.space}{A character string specifies one of the four sides,
which can be one of \code{'top'}, \code{'bottom'}, \code{'left'} and \code{'right'}. Default,
\code{legend.space = 'top'}.}

\item{legend.ncol}{An integer specifies a number of columns, possibly
divided into blocks, each containing some rows. Default,
\code{legend.ncol = 2}.}

\item{...}{Other arguments to be passed to \code{\link[lattice]{xyplot}}.}
}
\description{
This function draws predicted curves (lines) against an explanatory variable
for each category of a categorical variable.
}
\examples{
\dontrun{
library(R2MLwiN)

## Read alevchem data
data(alevchem)

alevchem['gcseav'] = double2singlePrecision(alevchem['gcse-tot'] / alevchem['gcse-no'] - 6)
alevchem['gcse^2'] = double2singlePrecision(alevchem['gcseav']^2)
alevchem['gcse^3'] = double2singlePrecision(alevchem['gcseav']^3)

## Example: A-level Chemistry
formula = a-point ~ (0|cons + gcseav + gcse^2 + gcse^3 + gender) + (1|cons)
levID = 'pupil'
estoptions = list(EstM = 1)
## Fit the model
mymodel = runMLwiN(formula, levID, indata = alevchem, estoptions = estoptions)

predCurves(mymodel, indata, xname = 'gcseav', group = 'gender')
}
}
\author{
Zhang, Z., Charlton, C.M.J., Parker, R.M.A., Leckie, G., and Browne,
W.J. (2015) Centre for Multilevel Modelling, University of Bristol.
}
\seealso{
\code{\link{predLines}}
}

