% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{print.mlwinfitIGLS}
\alias{print.mlwinfitIGLS}
\title{Summarize "mlwinfitIGLS" objects}
\usage{
\method{print}{mlwinfitIGLS}(x, ...)
}
\arguments{
\item{...}{other parameters}

\item{object, x}{an \code{\link{mlwinfitIGLS-class}} object}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If TRUE, 'significance stars' are printed for each coefficient.}
}
\description{
Summarize "mlwinfitIGLS" objects
}
\seealso{
\code{\link[base]{print}}
}
