% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{show,mlwinfitMCMC-method}
\alias{show,mlwinfitMCMC-method}
\title{Show objects of class "mlwinfitMCMC"}
\usage{
\S4method{show}{mlwinfitMCMC}(object)
}
\description{
Show objects of class "mlwinfitMCMC"
}
\seealso{
\code{\link[stats4]{show-methods}}
}
