% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{show.mlwinfitIGLS}
\alias{show.mlwinfitIGLS}
\title{Summarize "mlwinfitIGLS" objects}
\usage{
\method{show}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{an \code{\link{mlwinfitIGLS-class}} object}

\item{...}{other parameters}
}
\description{
Summarize "mlwinfitIGLS" objects
}
\seealso{
\code{\link[methods]{show}}
}
