% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\docType{methods}
\name{summary,mlwinfitIGLS-method}
\alias{summary,mlwinfitIGLS-method}
\alias{print,mlwinfitIGLS-method}
\alias{show,mlwinfitIGLS-method}
\title{Summarize "mlwinfitIGLS" objects}
\usage{
\S4method{summary}{mlwinfitIGLS}(object, ...)

\S4method{print}{mlwinfitIGLS}(x, digits = max(3, getOption("digits") -
  2), signif.stars = getOption("show.signif.stars"), ...)

\S4method{show}{mlwinfitIGLS}(object)
}
\arguments{
\item{object, x}{an \code{\link{mlwinfitIGLS-class}} object}

\item{...}{other parameters}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If TRUE, 'significance stars' are printed for each coefficient.}
}
\description{
Summarize "mlwinfitIGLS" objects
}
