% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\docType{methods}
\name{summary,mlwinfitMCMC-method}
\alias{summary,mlwinfitMCMC-method}
\alias{print,mlwinfitMCMC-method}
\alias{show,mlwinfitMCMC-method}
\title{Summarize "mlwinfitMCMC" objects}
\usage{
\S4method{summary}{mlwinfitMCMC}(object, ...)

\S4method{print}{mlwinfitMCMC}(x, digits = max(3, getOption("digits") -
  2), signif.stars = getOption("show.signif.stars"), z.ratio = TRUE,
  ...)

\S4method{show}{mlwinfitMCMC}(object)
}
\arguments{
\item{object, x}{an \code{\link{mlwinfitMCMC-class}} object}

\item{...}{other parameters}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If TRUE, 'significance stars' are printed for each coefficient.}

\item{z.ratio}{logical. If TRUE, z-ratio values are displayed for each coefficient.}
}
\description{
Summarize "mlwinfitMCMC" objects
}
