% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{summary,mlwinfitIGLS-method}
\alias{summary,mlwinfitIGLS-method}
\title{Summarize "mlwinfitIGLS" objects}
\usage{
\S4method{summary}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{an \code{\link{mlwinfitIGLS-class}} object}

\item{...}{other parameters}
}
\description{
Summarize "mlwinfitIGLS" objects
}
\seealso{
\code{\link[stats4]{summary-methods}}
}
