% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{summary,mlwinfitMCMC-method}
\alias{summary,mlwinfitMCMC-method}
\title{Summarize "mlwinfitMCMC" objects}
\usage{
\S4method{summary}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{an \code{\link{mlwinfitMCMC-class}} object}

\item{...}{other parameters}
}
\description{
Summarize "mlwinfitMCMC" objects
}
\seealso{
\code{\link[stats4]{summary-methods}}
}
