% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trajectories.R
\name{trajectories}
\alias{trajectories}
\title{Plots MCMC chain trajectories}
\usage{
trajectories(object, Range = c(1, 5000), selected = NULL)
}
\arguments{
\item{object}{A data frame of parameter chains from MLwiN.}

\item{Range}{An integer vector of length two specifying the first and last
iterations of the chains.}

\item{selected}{A character vector specifying the selected chains to be
plotted.}
}
\description{
This function draws trajectories of MCMC chains.
}
\examples{
\dontrun{
library(R2MLwiN)

## Example: tutorial
data(tutorial, package = "R2MLwiN")

(mymodel <- runMLwiN(normexam ~ 1 + standlrt + (student | 1 + standlrt), estoptions = list(EstM = 1), data = tutorial))

trajectories(mymodel, Range = c(4501, 5000))
}
}
\author{
Zhang, Z., Charlton, C.M.J., Parker, R.M.A., Leckie, G., and Browne,
W.J. (2015) Centre for Multilevel Modelling, University of Bristol.
}
\seealso{
\code{\link{sixway}}
}

