% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitIGLS.R
\name{update.mlwinfitIGLS}
\alias{update.mlwinfitIGLS}
\title{Update "mlwinfitIGLS" objects}
\usage{
\method{update}{mlwinfitIGLS}(object, ...)
}
\arguments{
\item{object}{a valid \code{mlwinfitIGLS} class object with an R function call component named \code{call}, the expression used to create itself.}

\item{...}{additional arguments to the call, or arguments with changed values.}
}
\value{
either a new updated \code{mlwinfitIGLS} class object, or else an unevaluated expression for creating such an object.
}
\description{
Update "mlwinfitIGLS" objects
}
\seealso{
\code{\link[stats]{update}}
}
