% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwinfitMCMC.R
\name{update.mlwinfitMCMC}
\alias{update.mlwinfitMCMC}
\title{Update "mlwinfitMCMC" objects}
\usage{
\method{update}{mlwinfitMCMC}(object, ...)
}
\arguments{
\item{object}{a valid \code{mlwinfitMCMC} class object with an R function call component named \code{call}, the expression used to create itself.}

\item{...}{additional arguments to the call, or arguments with changed values.}
}
\value{
either a new updated \code{mlwinfitMCMC} class object, or else an unevaluated expression for creating such an object.
}
\description{
Update "mlwinfitMCMC" objects
}
\seealso{
\code{\link[stats]{update}}
}
